requirements.txt
asgiref==3.3.4
astroid==2.5.6
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==3.2.4
django-bs4==0.0.4
django-import-export==3.2.0
django-summernote==0.8.20.0
django-tinymce==3.6.1
et-xmlfile==1.1.0
filelock==3.12.0
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
import-export==0.3.1
isort==5.8.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
platformdirs==3.5.1
pylint==2.8.3
python-dateutil==1.5
pytz==2021.1
PyYAML==6.0
requests==2.25.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.1
tablib==3.5.0
toml==0.10.2
tzdata==2023.3
urllib3==1.26.5
virtualenv==20.23.0
waitress==2.0.0
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
